Moved GET /packages to POST /packages/install
A considerable amount of stuff was changed this commit. /GET packages
used to allow you to choose between if you wanted files or metadata. However,
this didn't work well with the fact that if you wanted to install a package
you also had to install its dependencies. Now, you can make a POST
to /packages/install with a body of all the packages you already have
installed and the lambda will figure out what new files you need and it will
serve them to you! Getting package metadata can now be reserved solely for
GET /packages, which is much more fitting.